Remove the configuration directory and move the API configuration file to the directory services
